Fold traffic/playbook selftests into pytest and drop stale _banner refs
gitm/_banner.py was deleted but traffic/playbook CLIs and a selftest still
imported it, failing CI. The checks now live in tests/ as plain test_*
functions, so --selftest, CHECKS and run_all are removed from the package.
